correccion de observaciones presentadas en despacho de servicios

+++ Fuentes/AHSECO.CCL.FRONTEND/AHSECO.CCL.FRONTEND/resources/js/app/ventas/solicitud.js
@@ -2364,6 +2364,16 @@
                 $cmbFlujo.val(codFlujo).trigger("change.select2");
             }
 
+            if (rol == "SGI_VENTA_COORDINASERV" || rol == "SGI_VENTA_COORDINAATC") {
+                if ($TipoSolicitud.val() === "TSOL01") {
+                    $('#cmbTipoDocumentoCarga option[value="DVT03"]').remove(); //Factura
+                    $('#cmbTipoDocumentoCarga option[value="DVT04"]').remove(); //Ficha Instalacion
+                    $('#cmbTipoDocumentoCarga option[value="DVT06"]').remove(); //Guia de BO
+                    $('#cmbTipoDocumentoCarga option[value="DVT07"]').remove(); //Guia de Pedidos
+                    $('#cmbTipoDocumentoCarga option[value="DVT08"]').remove(); //Guia de Remision
+                }         
+            }
+
 
             if ($numeroSolicitud.val() != "") {
 
@@ -2407,11 +2417,16 @@
                     $openRegdateOrdenCompra.prop('disabled', true);
                 }
 
+                if ($idRolUsuario.val() === "SGI_VENTA_FACTURA") {
+                    $btnAgregarDocumento.hide();
+                }
 
-                if ($estadoSol.val() == "SFIN" || $estadoSol.val() == "NOVT") {
+                if (($estadoSol.val() == "SFIN" || $estadoSol.val() == "NOVT" || $estadoSol.val() =="VTPG")) {
                     $btnCargarDocumento.hide();
                     $btnAgregarObservacion.hide();
                     $btnAgregarDocumento.hide();
+
+                    
                 }
 
                 if (data.Result.ContadorCabecera.ContadorSinStock > 0) {
@@ -2608,7 +2623,7 @@
                         var html = '<div class="text-center">';
                         //var d = "'" + data.Result.Adjuntos[i].CodigoDocumento + "','" + data.Result.Adjuntos[i].RutaDocumento + "'";
                         html += ' <a class="btn btn-default btn-xs" title="Descargar"  href="javascript:solicitud.download(' + data.Result.Adjuntos[i].CodigoDocumento + ')"><i class="fa fa-download" aria-hidden="true"></i></a>&nbsp;';
-                        if ($estadoSol.val() != "SFIN" && $estadoSol.val() != "NOVT") {
+                        if (($estadoSol.val() != "SFIN" && $estadoSol.val() != "NOVT" && $estadoSol.val() != "VTPG") && $idRolUsuario.val() != "SGI_VENTA_FACTURA") {
                             html += ' <a class="btn btn-default btn-xs" title="Eliminar"  href="javascript:solicitud.eliminarDocumento(' + data.Result.Adjuntos[i].CodigoDocumento + ')"><i class="fa fa-ban" aria-hidden="true"></i></a>&nbsp;';
                         }
                         html += '</div>';
@@ -2868,7 +2883,7 @@
                                 }                             
                             );
                             var html = '<div class="text-center">';
-                            if ($estadoSol.val() != "SFIN" && $estadoSol.val() != "NOVT") {
+                            if (($estadoSol.val() != "SFIN" && $estadoSol.val() != "NOVT" && $estadoSol.val() != "VTPG") && $idRolUsuario.val() != "SGI_VENTA_FACTURA") {
                                 html += ' <a class="btn btn-default btn-xs" title="Eliminar"  href="javascript:solicitud.eliminarDocumento(' + data.Result.Codigo + ')"><i class="fa fa-ban" aria-hidden="true"></i></a>';
                             }
                             html += ' <a class="btn btn-default btn-xs" title="Descargar"  href="javascript:solicitud.download(' + data.Result.Codigo + ')"><i class="fa fa-download" aria-hidden="true"></i></a>&nbsp;';
@@ -3559,7 +3574,9 @@
             ext == "xls" || ext == "XLS" ||
             ext == "xlsx" || ext == "XLSX" ||
             ext == "doc" || ext == "DOC" ||
-            ext == "docx" || ext == "DOCX") {
+            ext == "docx" || ext == "DOCX" ||
+            ext == "zip" || ext == "ZIP" ||
+            ext == "rar" || ext == "RAR") {
             //beforeSendCargaDoc();
             var formdata = new FormData(); //FormData object
             //Appending each file to FormData object
